requirements.txt
ffmpeg-python==0.2.0
faster-whisper